+++ CHANGELOG.md
@@ -4,10 +4,12 @@ FEATURES:
 
 * **New Resource** `hostname`
 
+
 IMPROVEMENTS:
 
 * errors: added additional context to error messages of pullzones and edgerules
 * resource/pullzone: new attribute: `cache_error_responses`
+* resource/pullzone: new block `safehop`
 
 ## 0.2.0 (November 12, 2021)